+++ src/Management/ThemePresetProviderInterface.php
@@ -1,0 +1,15 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+namespace c975L\ConfigBundle\Management;
+
+interface ThemePresetProviderInterface
+{
+    // Each entry: id => ['label' => string, 'values' => [slug => value]] (slugs match Config::GROUP_THEME)
+    public function getPresets(): array;
+}